Delete old source gen test script only keeping debug print method
